requirements.txt

certifi==2023.5.7
charset-normalizer==3.1.0
configobj==5.0.8
pip==19.2.3
psycopg2-binary~=2.9.6
pymcprotocol==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-snap7==1.3
pytz-deprecation-shim==0.1.0.post0
pytz==2020.5
requests==2.28.2
schedule==1.1.0
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
zope.interface==6.0
